[/] [SetPixels] Try to read all announced pixels range..
Before it was

  // [1]
  Start = GetInt()
  Stop = GetInt()
  for Index in (Start, Stop)
    Red, Green, Blue = GetInt(), GetInt(), GetInt()

    if not (Red and Green and Blue)
      return

    // [2]
    if not SetPixel(Index, Red, Green, Blue)
      return

So when client thinks we have 100 pixels (and we have 60) and sends us
300 integers we were returning after reading 180 integers.

Other 120 integers still needs to be consumed by someone. I guess
it was [me_Menu] command getter. Which tried to find command for
each of them.

Now it's same (1) but (2) is

  SetPixel(Index, Red, Green, Blue)

We don't care that we asking driver to write past our stripe.
Driver can handle it.

We care that if it was announced that 40k values will be send,
we'll consume all that 40k values.